# openapi.yaml
openapi: 3.0.3
info:
  title: Resort Backend API (compat + core)
  version: 1.0.0
  description: "Machine-readable contract for the Resort Backend used by the frontend. Includes compat `/api` endpoints, admin/internal headers, and upload specs."
servers:
  - url: "{API_BASE}"
    description: "Runtime server (use VITE_API_URL or REACT_APP_API_URL as API_BASE)"
    variables:
      API_BASE:
        default: "http://localhost:8000"
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  schemas:
...
    GalleryItem:
      type: object
      properties:
        id: { type: string }
        filename: { type: string }
        url: { type: string }
        category: { type: string }
        visible: { type: boolean }
paths:
...
  /packages/:
    get:
      summary: List packages
      responses:
        '200': { description: OK }
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
  /packages/{package_id}:
    get:
      parameters: [{ name: package_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      parameters: [{ name: package_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: package_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
...
  /programs/:
    get:
      summary: List programs
      parameters:
        - name: tag
          in: query
          schema: { type: string }
      responses: { '200': { description: OK } }
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
  /programs/{program_id}:
    get:
      parameters: [{ name: program_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      parameters: [{ name: program_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: program_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
  /programs/recommend:
    post:
      summary: Recommend programs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                guestPreferences: { type: object }
                stayDays: { type: integer }
      responses: { '200': { description: OK } }
  /navigation/:
    get:
      summary: Get navigation (public fallback supported)
      parameters: [{ name: public, in: query, schema: { type: boolean } }]
      responses: { '200': { description: OK } }
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
  /navigation/{item_id}:
    put:
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
...
  /gallery/:
    get:
      summary: List gallery
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: visible
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer }
        - name: skip
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Paged gallery
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/GalleryItem'
                  count: { type: integer }
    post:
      summary: Upload gallery item (admin)
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                imageUrl:
                  type: string
      responses:
        '201': { description: Created }
  /gallery/{item_id}:
    get:
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      security: [{ AdminKey: [] }]
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      security: [{ AdminKey: [] }]
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
...
  /api/cottages:
    get:
      summary: Compat cottages endpoint
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses: { '200': { description: OK } }
  /api/cottages/{cottage_id}:
    get:
      parameters: [{ name: cottage_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
...
security: []